SDRAM Reconfigure: automatically detect interface for programming.
The microblaze can now detect which interface the SDRAM reconfigure
command is being sent over and then sets the programming mux in the FPGA
to select the relevant interface.
The register to set the mux is at offset 0x18 and the following bits are
of interest for this command:
bit #1 -> USER (=1) or CONFIG (=0) data
bit #2 -> 1gbe i/f (=1) or 40gbe i/f (=0)
The SKARAB can now be programmed with casperfpga over both the 1gbe or
40gbe regardless of the other link's status (previously the 40gbe i/f
got priority if it's link was up).