Changes to eth mac host read and transmit functionality.
Peralex:
"Fixed bug in the MAC read host packet where it was possible to exceed
the packet buffer size. Packet buffer size increased to match actual
size of host packet buffer in the firmware. Check for overflow also
fixed.There was a bug in the "ReadHostPacket" function in "eth_mac.c".
The checking of the received host packet size against the "puReceivePacket"
buffer size was not being done correctly, therefore creating a risk of
overwriting memory locations outside the memory allocated for the
"puReceivePacket" buffer. Also, the size of "puReceivePacket" was 1KB
but the buffer in "wishbone_ten_gb_eth_attach" component is able to
capture packets up to 2KB in size. The size of "puReceivePacket" has
been increased to 2KB and the checking of the received host packet size
has been fixed."