Added UpdateGBEPHYConfiguration function.
Peralex:
This method updates the configuration of the 1GBE PHY to improve the
link compatibility with different NICs. The drivers for some NICs do not
support sleep reliably resulting in occasional packet loss. Also enabled
flow control through pause frames to prevent possible packet loss in the
Marvell 1GBE PHY. This is equivalent to setting DIS_SLEEP = '1' and
ENA_PAUSE = '1'.