DHCP - bypass the RENEW state and go back to the DISCOVER state.
- This circumvents a specific case where a relay agent appends tags to
the dhcp messages during lease acquisition and an IP is allocated by the
server based on these tags. Upon lease renewal, the client unicasts a
renew request directly to the server (in line with RFC 2131) thus
bypassing the relay agent. The tags are thus not appended to the
message. This appears to result in the server not renewing the lease. To
get around this, we will return to the DISCOVER state instead and re-acquire
the lease.

Note: Lease renewal is still available by disabling the flag
*flagDHCP_SM_SHORT_CIRCUIT_RENEW*.