Bug fix in main.c
The original Peralex code has a for-loop which cycles through the
interfaces to perform certain tasks on each interface. This bug fix
re-ordered the LLDP and DHCP tasks in main() and placed them outside
of this for-loop since these have their own for-loops which cycles
through the interfaces. This may have possibly interfered with some
logic.