Added optional time profiling code to debug programming latency.
In an attempt to debug control message latencies, more specifically the
latencies during SDRAM programming with the command protocol, code was
added to profile the time from when the cpu reads the RX buffer till
when it responds via the TX buffer.
This code is guarded by the TIME_PROFILE preprocessor macro and when
enabled, will significantly decrease programming performance since it
writes out debug time data to the serial port.
TODO: Further work to be done: do even finer time profiling in order to
determine where most of the latency is being added.