Reconfigure FPGA upon HMC init failure (hard-coded WB addr offsets)
[Built to test SKARAB HMC recovery after init failure.]

Read back the HMC init bits for mezz 0,1 and 2 and reconfigure the FPGA
if all the HMCs are not initialized within 2 seconds. These addresses
are currently set in the toolflow and have been hard-coded in this commit
to match a specific toolflow build. This may need to be reworked in the
design to be located at a fixed known wishbone address offset.

Build dependency:
-This commit will only work if **all three HMCs** are populated and
 compiled into the design otherwise the SKARAB will enter a continuous
 reconfiguration loop since it is designed to check for all three HMCs.

-WB address offsets for HMC status registers:
 hmc0 @ 0x700c0
 hmc1 @ 0x7005c
 hmc2 @ 0x70058