HMC: Implemented software to support the addition of MEZZ status register
The mezzanine status register C_RD_MEZZANINE_STAT_1_ADDR has been added
in firmware and is read by the Microblaze to determine whether an HMC
card(s) is present and in turn, determine when it has finished POST-ing
and initializing. A 10s time out is allowed for this and if any HMC does
not init within this time, an FPGA reconfigure from SDRAM is invoked.