DHCP: wait between 0 and 3 seconds before starting DHCP.
The waiting time is very a simple division "hash" of the mac address.
This was implemented in an attempt to offset and spread the DHCP
requests being sent by the SKARABs since they are all powered up or
reset at the same time. The three second upper time bound was
arbitrarily chosen and the 100ms Microblaze interrupt thus allows for
1 of 30 time slots to be allocated by the hash function (in reality,
only 29 time slots due to prime modulo operation in hash function).