WriteI2 cmd: now accepts 33 data bytes (no longer 32) in the payload.
This change was done to make provision for writing to the fan controller
IC via i2c. Some of its registers, the control lookup table in
particular, requires a block of 32 bytes of data when writing to it.
The WriteI2C command must thus make provision for a PMBus command byte
plus 32 bytes of data in the packet payload (thus 33 bytes).