Added persistent memory functionality and HMC reconfigure count.
Memory with persistent state between FPGA reconfigures is required
to hold programtic state upon a reconfigure.
This is used to hold the HMC-initiated reconfigure count and limit it to
prevent the board going into an endless reconfigure loop, resulting in
no remote access to the SKARAB.

For the persisten memory:
The fan controller MAX31785 chip on the SKARAB motherboard has RAM
which can be accessed and used for this purpose. This is done by writing
to the chip's registers (one which does not influence the chip operation).
For this purpose the registers associated with the MFR_LOCATION is
chosen, which gives us 8 bytes of memory to use.
When the FPGA reconfigures, this register can be read/written to and the
state will persist between reconfigures.

The 8 bytes of memory will be allocated as follows:
 * byte[0]    => counter for reconfigurations caused by HMC timeouts
 * byte[1]    => counter for reconfigurations caused by DHCP timeouts
 * byte[2-8]  => reserved for future use