DHCP - request previously held lease/IP first between reconfigurations.
Attempt to request the previously held IP address from the DHCP server
at startup after a reconfiguration. This is in line with the INIT-REBOOT
state in the RFC 2131. The maximum amount requests can be set before the
DHCP state machine moves to the discovery step. The IP address used in
the requests was cached in persistent memory from the previous
successful lease acquisition (Note: this is cleared on a power cycle).
The implementation is currently site-specific since this feature is only
available on the one 40gbe link that is connected up on site (Note:
expansion of this would require more persistent memory registers to be
implemented).