Configure the interface parameters with previous lease prior to DHCP.
Set the fabric interface parameters "preemptively" with the knowledge
that we will always receive the same DHCP IP lease for the same switch
interface on site. This allows us to communicate with the SKARAB before
DHCP actually completes. These settings only persist between "warm"
reconfigures and will thus not have an impact after power-down.
This feature was added in an attempt to overcome the DHCP "bottleneck"
issue currently experienced on site. This is a temporary solution and
should be removed in the future when the DHCP lease acquisition times
improve. This feature is also guarded at compile-time by the
PREEMPT_CONFIGURE_FABRIC_IF macro.
Use of this feature together with the RECONFIG_UPON_NO_DHCP feature
is strongly discouraged.