Added the link monitoring feature.
This feature replaces the deprecated reconfig-upon-no-DHCP feature. This
is only implemented on interface with id=1 since this is a site-specific
feature and the SKARABs on site only have one 40gbe link.
This mechanism monitors the rx activity on the link and times out after
{t} seconds (where {t} is stored in SKARAB motherboard EEPROM, MIN=30s).
Once the link monitor times out, the fpga is reconfigured on the premise
that the link did not init properly.