HMC stats: always write hmc_total_init_retries_recorded to flash after the hmc-bringup mechanism fails completely.
The logic has been changed to always write the
'hmc_total_init_retries_recorded' count (previously called
'hmc_total_retries') to the flash of all hmc cards, regardless of
whether that hmc failed in that round of fpga re-configurations or not.
Previously this was only recorded if an hmc was the cause of one of the
initialisations failing in a given round of board bring-ups.
[ JIRA ticket ST-210 ]