Added delay to all MAX31785 i2c read transactions
In an attempt to solve the issue where the fan controller device
(max31785) "locks up" and occasionally causes skarab shutdowns when the
device is continuously polled for the sensor data, a delay
was added before all i2c read transactions addressing this max31785 ic.
The delay of 1.4ms was obtained from the 'bus-free time' spec in the
datasheet.