Release v3.20.36 (NOTE: See new version number scheme in this comment)
Major changes since v3.14.4:

- IGMP - refactored the IGMP code.
- Improvements to the igmp state machine
- Added a log-select option for igmp
- Log-level is retained across resets and reconfigures
- CLI - removed parsing of '.' in idle state
- Handle QSFP+ *PHY* Mezzanine cards
- Added the multicast_leave_group command (cmd id 0x61)
- Added delay to all MAX31785 i2c read transactions
- Increased the MAX31785 i2c read delay to 5ms
- Added ?whoami cmd to cli to identify skarab via console
- Added ?uname cmd to cli
- Reduced minimum dhcp monitor timeout to 5s with 1s increase on each reset
- Added GET_DHCP_MONITOR_TIMEOUT command (ID 0x63)

Note: the software version number will now indicate the build date of
the elf as follows: v3.year.day_of_year
i.e. this version will be numbered as follows:
v3.20.36 - the major number 'v3' will be retained for SARAO releases, as
opposed to 'v2' for Peralex releases, '20' for the year 2020
and '36' for the 36th day of the year.