eth-mac - guard all trace level logging with additional check (programming speedup)
The trace level logging in the eth mac driver is bundled together and
guarded by an additional if-statement. This speeds up the normal
operation of the driver which is especially important for programming.
Each log_printf() incurs a cost to check the log-level and by bundling
these together with a higher-level check, it saves cpu cycles during
normal operation.