eth_mac - write debug value to respective core's ETH_MAC_REG_CNT_RESET reg (upon TX only)
This is useful during debugging and as a sanity check to ensure the
wishbone to eth ID mapping is correct. This logic writes the respective
core WB-offset OR'd with the physical interface ID to the respective
register. This is only invoked when the log level is set to trace. It is
also not available in the legacy WB mapping since the register set
differs.  The following is written to the respective register:
0x84001 (=0x84000 | 1) for core0...
0x9A002 (=0x9A000 | 2) for core1...
0xB0003 for core2... etc